add french symbols and shit
