+++ keyboards/crkbd/keymaps/my_keymap/helpers.c
@@ -1,14 +1,10 @@
-// +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+#include "quantum.h"
+#include "keymap.h"
+
 bool is_accent_layer(void) {
-  return layer_state_is(_ACCENTS);
-  or:
-  return IS_LAYER_ON(_ACCENTS); // which is just an alias
+  return IS_LAYER_ON(_ACCENTS);
 }
 
 bool is_shift_layer(void) {
-  return layer_state_is(_SFT_COLEMAK_FR);
-  or:
-  return IS_LAYER_ON(_SFT_COLEMAK_FR); // which is just an alias
+  return IS_LAYER_ON(_SFT_COLEMAK_FR);
 }
-
-// +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
