+++ keyboards/crkbd/keymaps/my_keymap/helpers.h
@@ -1,0 +1,4 @@
+#pragma once
+
+bool is_accent_layer(void);
+bool is_shift_layer(void);
